pyproject.toml
[tool.poetry]
name = "exam2pptvideo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.0"
python-pptx = "^0.6.18"
pdf2image = "^1.15.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
Sphinx = "^4.0.1"
sphinx-rtd-theme = "^0.5.2"

[tool.poetry.scripts]
exam_pptx_validate='exam2pptvideo._entry.validation:validate'
exam_csv2pptx='exam2pptvideo._entry.command:csv2pptx'
exam_pptx2media='exam2pptvideo._entry.command:pptx2pdf2images'
exam_pptx2video='exam2pptvideo._entry.command:pptx2video'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
